+++ libraries/pipes/implementations/rust/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- (pull/71) Added the `close` method to `PipesContext`
 - (pull/67) Added the `data_version` parameter to `report_asset_materialization`
 - (pull/60) Added `AssetCheckSeverity` to the jsonschema definitions
 - (pull/59) Moved dagster pipes version into a constant